web_app/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
python-dotenv>=1.0.0
# Authentication
PyJWT>=2.8.0
cryptography>=41.0.0
# Reuse existing CLI dependencies
anthropic
openai
requests
pyyaml